+++ dakara_server/playlist/permissions.py
@@ -4,7 +4,7 @@ from django.contrib.auth import get_user_model
 
 from internal.permissions import BasePermissionCustom
 from library.models import Song
-from playlist.models import PlaylistEntry, Karaoke
+from playlist.models import PlaylistEntry
 
 
 UserModel = get_user_model()
@@ -75,15 +75,3 @@ class IsSongEnabled(BasePermissionCustom):
 
         except ObjectDoesNotExist:
             return True
-
-
-class KaraokeIsNotStoppedOrReadOnly(permissions.BasePermission):
-    """Grant access to not safe views if the kara is not in stop mode
-    """
-
-    def has_permission(self, request, view):
-        if request.method in permissions.SAFE_METHODS:
-            return True
-
-        karaoke = Karaoke.get_object()
-        return karaoke.status != Karaoke.STOP