+++ dakara_server/playlist/views.py
@@ -3,6 +3,7 @@ from datetime import datetime, timedelta
 
 from django.utils import timezone
 from django.conf import settings
+from django.core.cache import cache
 from django.shortcuts import get_object_or_404
 from django.contrib.auth import get_user_model
 from rest_framework import status
@@ -19,6 +20,7 @@ from playlist import models
 from playlist import serializers
 from playlist import permissions
 from playlist.consumers import broadcast_to_channel
+from playlist.date_stop import KARAOKE_JOB_NAME, scheduler, clear_date_stop
 
 tz = timezone.get_default_timezone()
 logger = logging.getLogger(__name__)
@@ -74,7 +76,6 @@ class PlaylistEntryListView(drf_generics.ListCreateAPIView):
         permissions.IsPlaylistUser | internal_permissions.IsReadOnly,
         (permissions.IsPlaylistManager & library_permissions.IsLibraryManager)
         | permissions.IsSongEnabled,
-        permissions.KaraokeIsNotStoppedOrReadOnly,
     ]
     queryset = models.PlaylistEntry.get_playlist()
 
@@ -99,6 +100,22 @@ class PlaylistEntryListView(drf_generics.ListCreateAPIView):
         return Response(serializer.data)
 
     def perform_create(self, serializer):
+        # Deny creation if kara is not ongoing
+        karaoke = models.Karaoke.get_object()
+        if not karaoke.ongoing:
+            raise PermissionDenied(detail="Karaoke is not ongoing.")
+
+        # Deny creation if karaoke does not allow new entries,
+        # and user is not manager.
+        if (
+            not karaoke.can_add_to_playlist
+            and not self.request.user.is_playlist_manager
+            and not self.request.user.is_superuser
+        ):
+            raise PermissionDenied(
+                detail="Karaoke was set to disallow playlist entries creation."
+            )
+
         # Deny the creation of a new playlist entry if it exceeds the playlist
         # capacity set in settings.
         queryset = self.filter_queryset(self.get_queryset())
@@ -112,7 +129,6 @@ class PlaylistEntryListView(drf_generics.ListCreateAPIView):
         # level, as permission examination takes place before the data are
         # validated. Moreover, the object permission method won't be called as
         # we are creating the object (which obviously doesn't exist yet).
-        karaoke = models.Karaoke.get_object()
 
         if karaoke.date_stop is not None and not (
             self.request.user.is_playlist_manager or self.request.user.is_superuser
@@ -147,15 +163,15 @@ class PlaylistEntryListView(drf_generics.ListCreateAPIView):
         # Request the player to play the latest playlist entry immediately if :
         #   - it exists;
         #   - the playlist was empty beforehand;
-        #   - the kara status is set to play;
+        #   - player is set to play next song
         #   - the player is idle.
         next_playlist_entry = models.PlaylistEntry.get_next()
         player = models.Player.get_or_create()
         if all(
             (
-                karaoke.status == models.Karaoke.PLAY,
-                playlist_was_empty,
                 next_playlist_entry is not None,
+                playlist_was_empty,
+                karaoke.player_play_next_song,
                 player.playlist_entry is None,
             )
         ):
@@ -184,16 +200,15 @@ class PlayerCommandView(drf_generics.UpdateAPIView):
         permissions.IsPlaylistManager
         | permissions.IsPlayingEntryOwner
         | internal_permissions.IsReadOnly,
-        permissions.KaraokeIsNotStoppedOrReadOnly,
     ]
     serializer_class = serializers.PlayerCommandSerializer
 
     def perform_update(self, serializer):
-        # check the karaoke is not stopped
+        # check the karaoke is ongoing
         karaoke = models.Karaoke.get_object()
-        if karaoke.status == models.Karaoke.STOP:
+        if not karaoke.ongoing:
             raise PermissionDenied(
-                "The player cannot receive commands if the " "karaoke is stopped"
+                "The player cannot receive commands if the karaoke is not ongoing"
             )
 
         # check the player is not idle
@@ -263,15 +278,34 @@ class KaraokeView(drf_generics.RetrieveUpdateAPIView):
         """Update the karaoke
         """
         super().perform_update(serializer)
-
-        # if the status of the karaoke has changed
-        if "status" not in serializer.validated_data:
-            return
-
-        # empty the playlist and clear the player if the new status is stop
         karaoke = serializer.instance
 
-        if karaoke.status == models.Karaoke.STOP:
+        # Management of date stop
+
+        if "date_stop" in serializer.validated_data:
+            # Clear existing scheduled task
+            existing_job_id = cache.get(KARAOKE_JOB_NAME)
+            if existing_job_id is not None:
+                existing_job = scheduler.get_job(existing_job_id)
+                if existing_job is not None:
+                    existing_job.remove()
+                    logger.debug("Existing date stop job was found and unscheduled")
+
+                else:
+                    cache.delete(KARAOKE_JOB_NAME)
+
+            if karaoke.date_stop is not None:
+                # Schedule date stop clear
+                job = scheduler.add_job(
+                    clear_date_stop, "date", run_date=karaoke.date_stop
+                )
+                cache.set(KARAOKE_JOB_NAME, job.id)
+                logger.debug("New date stop job was scheduled")
+
+        # Management of kara status Booleans change
+
+        # empty the playlist and clear the player if the kara is switched to not ongoing
+        if "ongoing" in serializer.validated_data and not karaoke.ongoing:
             # request the player to be idle
             broadcast_to_channel("playlist.device", "send_idle")
 
@@ -285,7 +319,13 @@ class KaraokeView(drf_generics.RetrieveUpdateAPIView):
             # empty the player errors
             models.PlayerError.objects.all().delete()
 
-        elif karaoke.status == models.Karaoke.PLAY:
+            return
+
+        if (
+            karaoke.ongoing
+            and "player_play_next_song" in serializer.validated_data
+            and karaoke.player_play_next_song
+        ):
             player = models.Player.get_or_create()
 
             # request the player to play the next song if idle,